+++ shipyard.py
@@ -76,9 +76,7 @@ _NODEPREP_FILE = ('shipyard_nodeprep.sh', 'scripts/shipyard_nodeprep.sh')
 _GLUSTERPREP_FILE = ('shipyard_glusterfs.sh', 'scripts/shipyard_glusterfs.sh')
 _HPNSSH_FILE = ('shipyard_hpnssh.sh', 'scripts/shipyard_hpnssh.sh')
 _JOBPREP_FILE = ('docker_jp_block.sh', 'scripts/docker_jp_block.sh')
-_BLOBINGRESS_FILE = (
-    'shipyard_blobingress.sh', 'scripts/shipyard_blobingress.sh'
-)
+_BLOBXFER_FILE = ('shipyard_blobxfer.sh', 'scripts/shipyard_blobxfer.sh')
 _CASCADE_FILE = ('cascade.py', 'cascade/cascade.py')
 _SETUP_PR_FILE = (
     'setup_private_registry.py', 'cascade/setup_private_registry.py'
@@ -422,7 +420,7 @@ def add_pool(batch_client, blob_client, config):
     except KeyError:
         prefix = None
     # create resource files list
-    _rflist = [_NODEPREP_FILE, _JOBPREP_FILE, _BLOBINGRESS_FILE, regfile]
+    _rflist = [_NODEPREP_FILE, _JOBPREP_FILE, _BLOBXFER_FILE, regfile]
     if not use_shipyard_docker_image:
         _rflist.append(_CASCADE_FILE)
         _rflist.append(_SETUP_PR_FILE)
@@ -489,7 +487,7 @@ def add_pool(batch_client, blob_client, config):
         pass
     # digest any input_data
     addlcmds = convoy.data.process_input_data(
-        blob_client, config, _BLOBINGRESS_FILE, config['pool_specification'])
+        config, _BLOBXFER_FILE, config['pool_specification'])
     if addlcmds is not None:
         start_task.append(addlcmds)
     del addlcmds
@@ -885,8 +883,7 @@ def main():
         convoy.batch.del_node(batch_client, config, args.nodeid)
     elif args.action == 'addjobs':
         convoy.batch.add_jobs(
-            batch_client, blob_client, config, _JOBPREP_FILE,
-            _BLOBINGRESS_FILE)
+            batch_client, blob_client, config, _JOBPREP_FILE, _BLOBXFER_FILE)
     elif args.action == 'cleanmijobs':
         convoy.batch.clean_mi_jobs(batch_client, config)
     elif args.action == 'termjobs':